@@ -1,28 +1,33 @@
-import React, { Suspense } from "react";
-import ReactDOM from "react-dom";
+import React from "react";
+import ReactDOM from "react-dom/client";
 import reportWebVitals from "./reportWebVitals";
-import { BrowserRouter } from "react-router-dom";
 import { Provider } from "react-redux";
-import Store from "./Redux/Store";
+import store from "./tools/redux/store";
+import App from "./App";
 
-// styles
 import "./index.css";
-import "antd/dist/antd.css";
-import "./Fonts/Font-stylesheet.css";
+import { QueryClient, QueryClientProvider } from "react-query";
 
-const App = React.lazy(() => import("./App"));
+const twentyFourHoursInMs = 1000 * 60 * 60 * 24;
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      refetchOnWindowFocus: false,
+      refetchOnmount: false,
+      refetchOnReconnect: false,
+      retry: false,
+      staleTime: twentyFourHoursInMs,
+    },
+  },
+});
 
-ReactDOM.render(
-  <React.StrictMode>
-    <BrowserRouter>
-      <Provider store={Store}>
-        <Suspense fallback={<div className={"h-screen"}></div>}>
-          <App />
-        </Suspense>
-      </Provider>
-    </BrowserRouter>
-  </React.StrictMode>,
-  document.getElementById("root")
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(
+  <Provider store={store}>
+    <QueryClientProvider client={queryClient}>
+      <App />
+    </QueryClientProvider>
+  </Provider>
 );
 
 // If you want to start measuring performance in your app, pass a function
