@@ -1,21 +1,40 @@
+/** @type {import('tailwindcss').Config} */
+
 module.exports = {
-  purge: ["./src/**/*.{js,jsx,ts,tsx}", "./public/index.html"],
-  darkMode: false, // or 'media' or 'class'
+  content: ["./src/**/*.{js,jsx,ts,tsx}"],
   theme: {
     extend: {
-      brightness: {
-        25: ".25",
-        175: "1.75",
+      colors: {
+        primary: "#161A27",
+        white: { main: "#E6E6E6", primary: "#DBE0F0" },
+        turquoise: { main: "#2ECBD1" },
+        pink: { main: "#F880C4" },
+        yellow: { main: "#F8DC68" },
+        blue: {
+          light: "#9FA9C6",
+          tint: "#8790AB",
+          gray: "#525C7A",
+          dark: "#1F2433",
+        },
       },
-      dropShadow: {
-        xl: "0 35px 35px rgba(0, 0, 0, 0.55)",
+      fontFamily: {
+        display: "Mulish",
       },
-    },
-  },
-  variants: {
-    extend: {
-      brightness: ["hover", "focus"],
-      filter: ["hover", "focus"],
+      spacing: {
+        21: "5.25rem",
+        22: "5.5rem",
+        30: "7.5rem",
+        34: "8.5rem",
+        66: "16.5rem",
+        68: "17rem",
+        86: "21.5rem",
+        98: "24.5rem",
+        112: "28rem",
+        119: "29.75rem",
+        124: "31rem",
+        128: "32rem",
+        132: "33rem",
+      },
     },
   },
   plugins: [],
